+++ src/test/scala/fr/janalyse/split/StringSplitTest.scala
@@ -4,9 +4,9 @@ import org.scalatest.FunSuite
 import org.scalatest.ShouldMatchers
 
 class StringSplitTest extends FunSuite with ShouldMatchers {
-  
+
   import StringSplit.split
-  
+
   test("Direct split tests") {
     split("") should equal(Vector())
     split("   ") should equal(Vector())
@@ -23,26 +23,24 @@ class StringSplitTest extends FunSuite with ShouldMatchers {
     split(""" a b,    c d """) should equal(Vector("a", "b,    c", "d"))
     split(""" a [b   c] d """) should equal(Vector("a", "b   c", "d"))
     split("""99.99.99.188 - - [14/Oct/2013:14:07:35 +0200] "GET /truc/ping.html HTTP/1.0" 200 7 - 1""") should equal(
-        Vector(
-            "99.99.99.188",
-            "-",
-            "-",
-            "14/Oct/2013:14:07:35 +0200",
-            "GET /truc/ping.html HTTP/1.0",
-            "200",
-            "7",
-            "-",
-            "1"
-            )
-        )
+      Vector(
+        "99.99.99.188",
+        "-",
+        "-",
+        "14/Oct/2013:14:07:35 +0200",
+        "GET /truc/ping.html HTTP/1.0",
+        "200",
+        "7",
+        "-",
+        "1"))
   }
-  
+
   test("extends split tests") {
     split("1 2 3", 2) should equal(Vector("1", "2 3"))
-    split("1 2 3", 1) should have size(1)
+    split("1 2 3", 1) should have size (1)
     split("1 [A B C] 3 4", 3) should equal(Vector("1", "A B C", "3 4"))
   }
-  
+
   test("split on log format specification") {
     val fmt = """%h %l %u %t "%r" %>s %b %D %{X-Forwarded-For}i %{sm_universalid}i "%{User-agent}i" %{msr_sequence_id}i"""
     val tokens = split(fmt)
@@ -56,19 +54,22 @@ class StringSplitTest extends FunSuite with ShouldMatchers {
     assert(tokens.size === 12, "")
     assert(tokens(4) === "%r")
   }
-  
+
   test("split test") {
-    val line="""truc [machin chose] 'gloubs' one, two, three, four 3454Mb long zut oups plouf paf le chien"""
-    def now = System.currentTimeMillis()
-    val started = now
-    var processedTokens = 0L
-    var processedLines=0L    
-    do {
-      val tokens = split(line)
-      processedTokens += tokens.size
-      processedLines += 1
-    } while (now - started < 10000L)
-    val howlong = (now - started)/1000
-    info(f"found ${processedTokens/howlong}%,d tokens/second through ${processedLines/howlong}%,d lines/second")
+    for { a <- 1 to 5 } {
+      val line = """truc [machin chose] 'gloubs' one, two, three, four 3454Mb long zut oups plouf paf le chien"""
+      def now = System.currentTimeMillis()
+      val started = now
+      var processedTokens = 0L
+      var processedLines = 0L
+      do {
+        val tokens = split(line)
+        processedTokens += tokens.size
+        processedLines += 1
+      } while (now - started < 5000L)
+      val howlong = (now - started) / 1000
+      info(f"found#$a : ${processedTokens / howlong}%,d tokens/second through ${processedLines / howlong}%,d lines/second")
+      Thread.sleep(1000)
+    }
   }
 }